webapp/requirements.txt
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.0.2
gunicorn==20.0.4
imagecodecs==2020.2.18
imageio==2.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==2.1.2
networkx==2.4
numpy==1.18.4
Pillow==5.2.0
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.4.1
six==1.14.0
tifffile==2020.5.11
torch==0.4.1
torchvision==0.2.1
Werkzeug==1.0.1
